backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
scikit-learn==1.3.2
numpy==1.26.2
pymongo==4.6.1
langchain==0.0.352
langchain-community==0.0.6
langchain-mongodb==0.0.1
python-dotenv==1.0.0
requests==2.31.0